backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
